Fix Linux build compatibility
- AITerminal: Replace CFAbsoluteTimeGetCurrent() with Date() (CoreFoundation unavailable on Linux)
- Highlighter: Replace Range(NSRange, in:) with cross-platform rangeFromNSRange helper (FoundationEssentials on Linux lacks this initializer)
- Terminal: Cast TIOCGWINSZ to UInt for ioctl (Int32 on Linux vs UInt on macOS)
- JSON: Use #if canImport(ObjectiveC) guard for CFGetTypeID/CFBooleanGetTypeID, fallback to objCType check on Linux
- Events: Use #if canImport(ObjectiveC) for objc_setAssociatedObject, fallback to dictionary-based storage on Linux
- Prompt: Use tcflag_t instead of UInt for termios c_lflag bitmask (UInt32 on Linux vs UInt on macOS)

Tested: macOS (swift build) and Linux (docker swift:6.0.3)

+++ Sources/RichSwift/Console/Terminal.swift
@@ -84,7 +84,7 @@ public struct Terminal: Sendable {
         return 80
         #else
         var size = winsize()
-        if ioctl(STDOUT_FILENO, TIOCGWINSZ, &size) == 0 {
+        if ioctl(STDOUT_FILENO, UInt(TIOCGWINSZ), &size) == 0 {
             return Int(size.ws_col)
         }
         
@@ -108,7 +108,7 @@ public struct Terminal: Sendable {
         return 24
         #else
         var size = winsize()
-        if ioctl(STDOUT_FILENO, TIOCGWINSZ, &size) == 0 {
+        if ioctl(STDOUT_FILENO, UInt(TIOCGWINSZ), &size) == 0 {
             return Int(size.ws_row)
         }
